Reload kernel wifi modules when network list is suspiciously empty
Issue:

Sometimes we need to reload the wifi kernel modules in order to make it
work again. I already do this automatically for connecting to a network.
Maybe it's okay to do it also when starting AP mode.

Solution:

To mitigate the issue, the kernel wifi module is reloaded if the network
list is empty suspiciously (3 times in a row). Since the frontend
refreshes automatically when there are no networks, it will try the
thing easily.